+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Removed the dependency on the `ffmpeg_kit` package.
 - Created `FFmpegConfig` to improve video and cover export functionality.
 - Added an example demonstrating how to export video and cover images using either `ffmpeg_kit` or `ffmpeg_wasm`, depending on the platform.
+- Added `defaultCoverThumbnailQuality` to `VideoEditorController.file` constructor.
 - Updated the `video_thumbnail` version to 0.6.0.
 - Updated the project to work on the web.
 - Updated the documentation to reflect these changes.